\name{SNP2GO-internal}
\alias{AddOffspringInformation}
\alias{ComputeGenomicRangesByLevel}
\alias{ComputeLevels}
\alias{ComputeSNPsByLevel}
\alias{FindLevel}
\alias{GetAllRegionsAndSNPs}
\alias{GetCandidateTerms}
\alias{GetProgressBarIncrementer}
\alias{GetStatistic}
\alias{KeepSNPs}
\alias{SamplingTest}
\alias{ProcessGFF}
\alias{ProcessGTF}
\alias{ProcessMartExport}
\alias{RemoveFromUpperLevels}
\title{Internal SNP2GO Functions}
\description{
          Internal SNP2GO functions
}
\details{
         These functions are not intended to be called by the user.
}
\keyword{ internal } 

