\name{SNP2GO-package}
\alias{SNP2GO-package}
\alias{SNP2GO}
\docType{package}
\title{
Package for the functional analysis of Genome Wide Association Studies
}
\description{
SNP2GO is an R package developed for the functional analysis of Genome-Wide Association Studies. SNP2GO takes the genome annotation and the sets of candidate and non-candidate SNPs as input. This information is integrated by the program to test whether a GO category has an overrepresentation of candidate SNPs. Afterwards, SNP2GO performs simulation analysis to test if the significance of a GO category is due to local effects, such as linkage disequilibrium.
}
\details{
The analysis is performed by calling the \code{snp2go} function.
\tabular{ll}{
Package: \tab SNP2GO\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-24\cr
License: \tab Artistic License 2.0\cr
}
}

\references{
D. Szkiba, M. Kapun, A. von Haeseler, and M. Gallach. 
SNP2GO: functional analysis of genome-wide association studies.
Genetics. In press
}

\author{
David Szkiba

Maintainer: david.szkiba(AT)univie.ac.at
}

\keyword{ package }

