% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp2go.R
\docType{package}
\name{SNP2GO-package}
\alias{SNP2GO}
\alias{SNP2GO-package}
\title{SNP2GO: Functional Analysis of Genome Wide Association Studies}
\description{
SNP2GO tests Gene Ontology terms for enrichment of candidate SNPs and infers whether enrichment is influenced by local effects.
}
\details{
The analysis is performed by calling the \code{snp2go} function.
}
\references{
D. Szkiba, M. Kapun, A. von Haeseler, and M. Gallach.
SNP2GO: functional analysis of genome-wide association studies.
(2014) Genetics. 197: 285-289 (DOI: 10.1534/genetics.113.160341)
}
\author{
David Szkiba
}
\keyword{package}
